# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Trà Đá Mentor Hub API",
    "description": "REST API for mentoring program operations. Authenticated routes require `Authorization: Bearer <accessToken>`. List endpoints return a raw array unless `page` is set, in which case the body is `{ data, page, limit, total }`. Errors use `{ success: false, code, message }`.",
    "version": "1.1.0",
    "contact": {
      "name": "Trà Đá Mentor",
      "url": "https://github.com/TheHien04/Tra-Da-Mentor-Hub"
    }
  },
  "servers": [
    { "url": "http://localhost:5000", "description": "Local development" },
    { "url": "/", "description": "Current host (production single-origin deploy)" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "Id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
      "Q": { "name": "q", "in": "query", "schema": { "type": "string" } },
      "Page": { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1 } },
      "Limit": { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100 } },
      "Track": { "name": "track", "in": "query", "schema": { "type": "string" } },
...
      "ApplicationStatus": { "name": "applicationStatus", "in": "query", "schema": { "type": "string" } },
...
    },
    "schemas": {
...
      "Mentee": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "userId": { "type": "string", "nullable": true },
          "name": { "type": "string" },
          "email": { "type": "string" },
          "phone": { "type": "string" },
          "track": { "type": "string" },
          "school": { "type": "string" },
          "interests": { "type": "array", "items": { "type": "string" } },
          "goals": { "type": "array", "items": { "type": "string" } },
          "progress": { "type": "number" },
          "mentorId": { "type": "string", "nullable": true },
          "groupId": { "type": "string", "nullable": true },
          "mentorshipType": { "type": "string", "enum": ["GROUP", "ONE_ON_ONE"] },
          "applicationStatus": { "type": "string" },
          "avatarUrl": { "type": "string" },
          "mentor": { "type": "string" },
          "group": { "type": "string" }
        }
      },
      "MenteeWrite": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string" },
          "phone": { "type": "string" },
          "track": { "type": "string" },
          "school": { "type": "string" },
          "interests": { "type": "array", "items": { "type": "string" } },
          "goals": { "type": "array", "items": { "type": "string" } },
          "progress": { "type": "number" },
          "mentorId": { "type": "string" },
          "groupId": { "type": "string" },
          "mentorshipType": { "type": "string" },
          "applicationStatus": { "type": "string" },
          "avatarUrl": { "type": "string" }
        }
      },
      "ApplicationStatusWrite": {
        "type": "object",
        "required": ["applicationStatus"],
        "properties": { "applicationStatus": { "type": "string" } }
      },
      "PagedMentees": {
        "type": "object",
        "required": ["data", "page", "limit", "total"],
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Mentee" } },
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" }
        }
      },
...
      "Notification": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "userId": { "type": "string" },
          "title": { "type": "string" },
          "message": { "type": "string" },
          "type": { "type": "string", "enum": ["info", "success", "warning", "error"] },
          "href": { "type": "string", "nullable": true },
          "read": { "type": "boolean" },
          "createdAt": { "type": "string" }
        }
      },
      "NotificationList": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Notification" } }
        }
      },
...
      "Ok": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "message": { "type": "string" }
        }
      }
    }
  },
  "paths": {
...
    "/api/mentors/{id}/mentees": {
      "get": {
        "tags": ["Mentors"],
        "summary": "Mentees assigned to a mentor",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "responses": {
          "200": {
            "description": "Mentees",
            "content": {
              "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Mentee" } } }
            }
          }
        }
      }
    },
...
    "/api/mentees": {
      "get": {
        "tags": ["Mentees"],
        "summary": "List mentees",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/Q" },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Limit" },
          { "$ref": "#/components/parameters/Track" },
          { "$ref": "#/components/parameters/ApplicationStatus" }
        ],
        "responses": {
          "200": {
            "description": "Mentees",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    { "type": "array", "items": { "$ref": "#/components/schemas/Mentee" } },
                    { "$ref": "#/components/schemas/PagedMentees" }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Mentees"],
        "summary": "Create mentee",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MenteeWrite" } } }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Mentee" } } }
          }
        }
      }
    },
    "/api/mentees/{id}": {
      "get": {
        "tags": ["Mentees"],
        "summary": "Get mentee",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "responses": {
          "200": {
            "description": "Mentee",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Mentee" } } }
          }
        }
      },
      "patch": {
        "tags": ["Mentees"],
        "summary": "Update mentee",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MenteeWrite" } } }
        },
        "responses": { "200": { "description": "Updated" } }
      },
      "delete": {
        "tags": ["Mentees"],
        "summary": "Delete mentee (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "responses": { "200": { "description": "Deleted" }, "403": { "description": "Forbidden" } }
      }
    },
    "/api/mentees/{id}/application-status": {
      "patch": {
        "tags": ["Mentees"],
        "summary": "Update application status",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApplicationStatusWrite" } } }
        },
        "responses": { "200": { "description": "Updated" } }
      }
    },
...
    "/api/groups/{id}/mentees": {
      "get": {
        "tags": ["Groups"],
        "summary": "Mentees in a group",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "responses": { "200": { "description": "Mentees" } }
      }
    },
    "/api/groups/{groupId}/mentees/{menteeId}": {
      "post": {
        "tags": ["Groups"],
        "summary": "Add a mentee to a group",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "groupId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "menteeId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Added" } }
      },
      "delete": {
        "tags": ["Groups"],
        "summary": "Remove a mentee from a group",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "groupId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "menteeId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Removed" } }
      }
    },
...
    "/api/notifications": {
      "get": {
        "tags": ["Notifications"],
        "summary": "Notifications for the signed-in user",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Notifications",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/NotificationList" } } }
          }
        }
      }
    },
    "/api/notifications/{id}/read": {
      "patch": {
        "tags": ["Notifications"],
        "summary": "Mark one notification read",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Id" }],
        "responses": { "200": { "description": "Updated" }, "404": { "description": "Not found" } }
      }
    },
    "/api/notifications/read-all": {
      "post": {
        "tags": ["Notifications"],
        "summary": "Mark every visible notification read",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Updated",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Ok" } } }
          }
        }
      }
    },
...
  }
}